frontend/requirements.txt
gradio
requests
python-decouple
